requirements.txt
asgiref==3.9.1
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
django-allauth==0.57.0
djangorestframework-simplejwt[rest_framework_simplejwt]==5.3.1
